fix(j-s): Renaming and cleanup
