+++ apps/system-e2e/src/tests/judicial-system/smoke/custody-tests.spec.ts
@@ -1,13 +1,14 @@
 import { expect } from '@playwright/test'
 import faker from 'faker'
-import { test } from '../utils/judicialSystemTest'
-import { randomPoliceCaseNumber, randomCourtCaseNumber } from '../utils/helpers'
 
 import { urls } from '../../../support/urls'
 import { verifyRequestCompletion } from '../../../support/api-tools'
-import { receiveAppealTest } from './shared/receive-appeal'
-import { appealCaseTest } from './shared/appeal-case'
-import { coaJudgeCompletesCaseTest } from './shared/coa-complates-case'
+import { test } from '../utils/judicialSystemTest'
+import { randomPoliceCaseNumber, randomCourtCaseNumber } from '../utils/helpers'
+
+import { receiveAppealTest } from './shared-steps/receive-appeal'
+import { appealCaseTest } from './shared-steps/send-appeal'
+import { coaJudgeCompletesCaseTest } from './shared-steps/complete-appeal'
 
 test.use({ baseURL: urls.judicialSystemBaseUrl })
