+++ apps/system-e2e/src/tests/judicial-system/smoke/shared-steps/complete-appeal.ts
